# Implement database optimizations