--- a/R/db.R
+++ b/R/db.R
@@ -508,9 +508,18 @@ collectResults <- function(expDir, dbName, dbDir = NULL, replace = FALSE,
 
   ## Remove database if replace is TRUE
   if (replace) {
-    file.remove(file.path(dbDir, paste0(dbName, ".sqlite")))
-    file.remove(file.path(dbDir, paste0(dbName, ".sqlite-shm")))
-    file.remove(file.path(dbDir, paste0(dbName, ".sqlite-wal")))
+    file.remove(
+      file.path(dbDir, paste0(dbName, ".sqlite")),
+      showWarnings = FALSE
+    )
+    file.remove(
+      file.path(dbDir, paste0(dbName, ".sqlite-shm")),
+      showWarnings = FALSE
+    )
+    file.remove(
+      file.path(dbDir, paste0(dbName, ".sqlite-wal")),
+      showWarnings = FALSE
+    )
   }
 
   ## Create connection