These improvements are supposed to make the queries to the database faster, especially when the results table is involved.

Also see #52