@rogoersTPH 2022-09-30
To add some numbers from benchmarks and tests:

Reading 250 outputs, each around 34k lines long. Then retrieving the results of 50 random scenarios from the database.

master branch DB creation:
```
##   user  system elapsed
## 48.086   0.531  20.952
## 337.5 MB DB size
```
master branch DB retrieval:
```
## 1701850 rows
##  user  system elapsed
## 2.203   0.059   2.267
```
PR branch DB creation (16 cores, batch strategy):
```
##   user  system elapsed
## 14.714   3.571  19.631
## 421.4 MB DB size
```
PR branch DB retrieval:
```
## 1701850 rows
##  user  system elapsed
## 1.484   0.038   1.529
```

The creation is not faster and the size is larger due to the index but retrieval is. The differences will get larger as the number of processed output files grow. 

Then some numbers from the new implementation, also using some of its features. 
```
fileCol <- function(scens, filter) {
  scens <- as.data.table(scens)
  filter <- substitute(filter)
  filesAggr <- scens[eval(filter), file]
  return(filesAggr)
}
...
fileFun = fileCol, fileFunArgs = list(scens = scens, filter = c(1:250)),
readFunArgs = list(filter = measure %in% c("nHost", "nSevere")), ...
```
This means we filter the input files on our own first (selecting the first 250 scenarios) and then extract only the rows for `nHost` and `nSevere`.
```
## 250 scens, 16 cores, batch
##  user  system elapsed
## 2.466   0.370   4.431

## 250 scens, 1 core, batch
##    user  system elapsed
##  44.477   0.510  12.128

## 250 scens, 16 core, serial
##   user  system elapsed
## 21.482   3.968  47.006

## 250 scens, 1 core, serial
##    user  system elapsed
## 49.067   0.649  14.152
```
The batch strategy with multiple cores is fastest (as expected) but interesting to see that multiple cores with the serial strategy is slowest. This should be due to the fact that SQLite with WAL allows multiple reads but only one a sinle write process. Thus, the muliple processes block each other. This strategy probably only makes sense, if the read and aggregation calculations take a longer time than adding the data to the data base.