@rogoersTPH 2022-09-30
Another addition, demonstrating the use of the injectable functions.

Premise: We have around 80k scenarios, the simulations are specific for a country district (called setting). In this example we are only interested in two specific districts and furthermore, we only care about the mean incidence rate and variance caused by the different seeds of the simulation (a strange example which doesn't make sense, but well .... ).

```r
fileCol <- function(scens, setting) {
  filesAggr <- scens[eval(parse(text = paste0("setting == \"", setting, "\""))), file]
  return(filesAggr)
}

aggrFoo <- function(df, scens.) {
  scens. <- as.data.table(scens.)
  scens. <- scens.[, scenario_id := ID]
  df <- df[, seed := scens[df, seed, on = "scenario_id"]]
  df <- df[measure %in% c("nUncomp", "nSevere", "nHost")]
  df <- dcast(df, scenario_id + seed + survey_date + third_dimension ~ measure, value.var = "value")
  df <- df[, IR := (nUncomp + nSevere) / nHost]
  df <- df[, list(mean_IR = mean(IR), sd_IR = sd(IR)), by = .(year(as.Date(survey_date)), third_dimension)]
  return(df)
}

system.time(
  for (s in c("maputocidade", "nampula")) {
    print(paste0("Working on ", s))
    collectResults(getCache("experimentDir"), "test_foo",
      replace = FALSE,
      fileFun = fileCol, fileFunArgs = list(scens = scens, setting = s),
      readFunArgs = list(filter = measure %in% c("nUncomp", "nSevere", "nHost")),
      aggrFun = aggrFoo, aggrFunArgs = list(scens. = scens),
      ncores = 16, strategy = "batch",
      resultsName = paste0("results_", s), resultsCols = list(
        names = c(
          "year", "third_dimension", "mean_IR", "sd_IR"
        ),
        types = c("INTEGER", "INTEGER", "NUMERIC", "NUMERIC")
      ),
      indexOn = list(c(paste0("results_", s), "year"))
    )
  }
)

## 11100 scenarios
##   user  system elapsed
## 79.240  23.913 125.015


con <- DBI::dbConnect(RSQLite::SQLite(), "test_foo.sqlite")
df1 <- as.data.table(DBI::dbReadTable(con, name = "results_maputocidade"))
DBI::dbDisconnect(con)
con <- DBI::dbConnect(RSQLite::SQLite(), "test_foo.sqlite")
df2 <- as.data.table(DBI::dbReadTable(con, name = "results_nampula"))
DBI::dbDisconnect(con)
df1 <- df1[, setting := "maputocidade"]
df2 <- df2[, setting := "nampula"]
df <- rbindlist(list(df1, df2))

## DB accesses are instantaneous

library(ggplot2)
ggplot(
  data = df[year > 2000],
  aes(
    x = year, y = mean_IR, color = as.factor(third_dimension),
    group = as.factor(third_dimension)
  )
) +
  geom_point() +
  geom_line() +
  geom_ribbon(
    aes(
      x = year, y = mean_IR, ymax = mean_IR + sd_IR, ymin = mean_IR - sd_IR,
      fill = as.factor(third_dimension)
    ),
    alpha = 0.2
  ) +
  facet_wrap(~setting) +
  xlab("Date") +
  ylab("Mean Incidence Rate") +
  labs(color = "Age group", fill = "Age group") +
  theme_bw()
```
![Rplot001](https://user-images.githubusercontent.com/91682637/193263262-6b295696-b927-44c6-afbf-993334746292.png)